Align indexing buffers and utility contracts
Preserve DocValues and norms state during flush, fix memory accounting,
and align radix-sort delegation and packed-value precision. Keep the
reviewed PointValuesWriter scope to accounting and length reporting.

Validation: cargo tidy and 43 targeted non-release regressions.

+++ src/core/util/packed/monotonic_block_packed_reader.rs
@@ -117,7 +117,7 @@ impl MonotonicBlockPackedReader {
   }
 }
 pub fn expected(origin: i64, average: f32, index: i32) -> i64 {
-  origin.wrapping_add(((average as f64) * (index as i64 as f64)) as i64)
+  origin.wrapping_add((average * index as f32) as i64)
 }
 
 pub struct MonotonicLongValues {
